+++ backend/src/main/java/com/coffee/backend/domain/auth/dto/KakaoUserInfoDto.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class KakaoUserInfoDto {
-    private Long userId;
+    private Long kakaoId;
 }
