+++ backend/src/main/java/com/coffee/backend/domain/user/repository/UserRepository.java
@@ -14,4 +14,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
     void deleteByUserUUID(String userUUID);
 
     Optional<User> findByEmail(final String email);
+
+    Optional<User> findByKakaoId(Long kakaoId);
 }
